damage post backend restriction

+++ backend/routes/api/damagePost.js
@@ -9,6 +9,7 @@ const DamagePost = require("../../models/damagePost.js")
 const Building = require("../../models/building.js")
 
 const requireAuth = require("../../middleware/requireAuth.js")
+const { canEditDamage } = require("../../util/permissionsCheckers.js")
 
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
@@ -57,55 +58,63 @@ async function getDamage(req, res, next) {
 
 //create one
 router.post("/", requireAuth, upload.any("images"), async (req, res) => {
-  let isBuildingExists = await Building.exists({ name: req.body.building })
+  if (canEditDamage(req.role)) {
+    let isBuildingExists = await Building.exists({ name: req.body.building })
 
-  if (req.body.sum) {
-    req.body.sum = Number(req.body.sum)
-  }
+    if (req.body.sum) {
+      req.body.sum = Number(req.body.sum)
+    }
 
-  let images = []
+    let images = []
 
-  if (req.files) {
-    images = req.files.map((file) => file.filename)
-  }
+    if (req.files) {
+      images = req.files.map((file) => file.filename)
+    }
 
-  if (isBuildingExists) {
-    const damage = new DamagePost({
-      name: req.body.name,
-      photo_links: images[0] != "" ? images : [],
-      description: req.body.description,
-      status: req.body.status,
-      building: req.body.building,
-      location: req.body.location,
-      sum: req.body.sum,
-    })
+    if (isBuildingExists) {
+      const damage = new DamagePost({
+        name: req.body.name,
+        photo_links: images[0] != "" ? images : [],
+        description: req.body.description,
+        status: req.body.status,
+        building: req.body.building,
+        location: req.body.location,
+        sum: req.body.sum,
+      })
 
-    try {
-      const newDamage = await damage.save()
+      try {
+        const newDamage = await damage.save()
 
-      res.status(201).json(newDamage)
-    } catch (err) {
-      res.status(400).json({ message: err.message })
+        res.status(201).json(newDamage)
+      } catch (err) {
+        res.status(400).json({ message: err.message })
+      }
+    } else {
+      res.status(400).json({ message: "Invalid building" })
     }
   } else {
-    res.status(400).json({ message: "Invalid building" })
+    res.status(500).json({ message: "Not enough rights" })
   }
 })
 
 // delete one
 router.delete("/:name", requireAuth, getDamage, async (req, res) => {
-  try {
-    res.damage.photo_links.forEach((link) => {
-      const address = path.resolve("./static/images/damage" + link)
-      if (fs.existsSync(address)) {
-        fs.unlinkSync(address)
-      }
-    })
+  if (canEditDamage(req.role)) {
+    try {
+      res.damage.photo_links.forEach((link) => {
+        const address = path.resolve("./static/images/damage" + link)
+        if (fs.existsSync(address)) {
+          fs.unlinkSync(address)
+        }
+      })
 
-    await res.damage.remove()
-    res.json({ message: "Deleted Damage post" })
-  } catch (err) {
-    res.status(500).json({ message: err.message })
+      await res.damage.remove()
+      res.json({ message: "Deleted Damage post" })
+    } catch (err) {
+      res.status(500).json({ message: err.message })
+    }
+  } else {
+    res.status(500).json({ message: "Not enough rights" })
   }
 })
 
@@ -116,59 +125,63 @@ router.patch(
   upload.any("images"),
   getDamage,
   async (req, res) => {
-    let isBuildingExists = await Building.exists({ name: req.body.building })
+    if (canEditDamage(req.role)) {
+      let isBuildingExists = await Building.exists({ name: req.body.building })
 
-    if (req.body.building != [] && req.body.faculty) {
-      isBuildingExists = await Building.exists({ name: req.body.building })
+      if (req.body.building != [] && req.body.faculty) {
+        isBuildingExists = await Building.exists({ name: req.body.building })
 
-      if (req.body.sum) {
-        req.body.sum = Number(req.body.sum)
-      }
-    }
-
-    let images = []
-
-    if (req.files) {
-      images = req.files.map((file) => file.filename)
-    }
-
-    res.damage.photo_links.forEach((link) => {
-      const address = path.resolve("./static/images/damage/" + link)
-      if (fs.existsSync(address)) {
-        fs.unlinkSync(address)
-      }
-    })
-    if (isBuildingExists) {
-      if (req.body.name != null) {
-        res.damage.name = req.body.name
-      }
-      if (req.body.photo_links != null) {
-        res.damage.photo_links = images
-      }
-      if (req.body.status != null) {
-        res.damage.status = req.body.status
-      }
-      if (req.body.building != null) {
-        res.damage.building = req.body.building
-      }
-      if (req.body.location != null) {
-        res.damage.location = req.body.location
-      }
-      if (req.body.description != null) {
-        res.damage.description = req.body.description
-      }
-      if (req.body.sum != null) {
-        res.damage.sum = req.body.sum
+        if (req.body.sum) {
+          req.body.sum = Number(req.body.sum)
+        }
       }
 
-      try {
-        const updatedBuilding = await res.damage.save()
-        res.json(updatedBuilding)
-      } catch (err) {
-        res.status(400).json({ message: err.message })
+      let images = []
+
+      if (req.files) {
+        images = req.files.map((file) => file.filename)
+      }
+
+      res.damage.photo_links.forEach((link) => {
+        const address = path.resolve("./static/images/damage/" + link)
+        if (fs.existsSync(address)) {
+          fs.unlinkSync(address)
+        }
+      })
+      if (isBuildingExists) {
+        if (req.body.name != null) {
+          res.damage.name = req.body.name
+        }
+        if (req.body.photo_links != null) {
+          res.damage.photo_links = images
+        }
+        if (req.body.status != null) {
+          res.damage.status = req.body.status
+        }
+        if (req.body.building != null) {
+          res.damage.building = req.body.building
+        }
+        if (req.body.location != null) {
+          res.damage.location = req.body.location
+        }
+        if (req.body.description != null) {
+          res.damage.description = req.body.description
+        }
+        if (req.body.sum != null) {
+          res.damage.sum = req.body.sum
+        }
+
+        try {
+          const updatedBuilding = await res.damage.save()
+          res.json(updatedBuilding)
+        } catch (err) {
+          res.status(400).json({ message: err.message })
+        }
+      } else {
+        res.status(400).json({ message: "Invalid building" })
       }
     } else {
-      res.status(400).json({ message: "Invalid building" })
+      res.status(500).json({ message: "Not enough rights" })
     }
   }
 )
